[MIRROR] Removes "jams"/un-used effect_transforms. (#10 code-quality/misc)
This is a [MIRROR] of Lambda-13/Civ13/pull/99.

This also removes some `effect_transforms`; which are redundant, and don't do anything.

Jams are a completely useless function, at least to my knowledge. It's supposed to prevent a barrage of bullets when lag occurs, but that rarely happens/will happen.

This is code-quality/misc.