Fully removes money/`real` amount sub-types. (#17)
* Full removal of money/real/five

Replaces it with a a variable type-ref declaration called `fiver`.

new()s it and equips it to the slots.

We do this because we cannot merge different paths and no one has set up the stacktype properly, or has checked if we have the type-path so we can properly check the amount. Removes bad code in Civ13.